final_pjt/final_pjt_final_ver2/back-server/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2022.6
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==2.0.2
